requirements.txt
aiogram
Pillow
python-dotenv